Split pytezos_core
