+++ pyproject.toml
@@ -20,17 +20,17 @@ classifiers = [
 ]
 packages = [
     { include = "pytezos", from = "src" },
+    { include = "pytezos_core", from = "src" },
     { include = "michelson_kernel", from = "src" },
 ]
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.11"
-attrs = "^21.4.0"
+attrs = "^22.1.0"
 base58 = "^2.1.1"
 bson = "^0.5.10"
 cattrs = "^22.1.0"
-# FIXME: Why?
-click = "^8.0.3,<=8.0.4"
+click = "^8.1.3"
 deprecation = "^2.1.0"
 docker = "^6.0.0"
 # FIXME: Broken `cattrs` dependency
@@ -79,7 +79,7 @@ sphinx-sitemap = "^2.2.0"
 types-python-dateutil = "^2.8.10"
 types-PyYAML = "^6.0.9"
 types-requests = "^2.26.3"
-types-setuptools = "^57.4.17"
+types-setuptools = "^65.3.0"
 types-simplejson = "^3.17.2"
 types-tabulate = "^0.8.4"
 parameterized = "^0.8.1"
